pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "molbind"
description = "A package for binding molecular modalities."
version = "0.0.1"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Adrian Mirza", email = "andrian.mirza@uni-jena.de" }]
keywords = ["embeddings", "contrastive learning"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: Unix",
  "Operating System :: MacOS",
]

requires-python = ">=3.11, <3.13"
dependencies = [
  "faiss-cpu",
  "torch",
  "torchvision",
  "torch_geometric",
  "numpy<2.0",
  "loguru",
  "transformers",
  "datasets==3.6.0",
  "fire",
  "rdkit",
  "pyarrow",
  "pandas",
  "polars",
  "info-nce-pytorch",
  "hydra-core==1.3.2",
  "fsspec",
  "lightning>=2.0.0",
  "torchmetrics>=0.11.4",
  "rootutils",
  "pre-commit",
  "class_resolver",
  "rich",
  "wandb",
  "hydra-colorlog==1.2.0",
  "hydra-optuna-sweeper==1.2.0",
  "omegaconf",
  "pyqt6>=6.9.1",
]

[project.optional-dependencies]
dev = [
  "codecov-cli>=0.4.1",
  "pytest>=7.4.0",
  "pytest-cov>=3.0.0",
  "ruff>=0.0.285",
]
docs = [
  "mkdocs-material>=9.4.0",
  "mkdocstrings[python]>=0.22.0",
  "mkdocs-gen-files>=0.5.0",
  "mkdocs-literate-nav>=0.6.0",
  "pillow>=10.0.0",
  "cairosvg>=2.7.1",
]

[tool.setuptools.package-data]
template = ["py.typed"]

[tool.pyright]
include = ["template"]
exclude = ["**/__pycache__"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-p no:warnings", "--import-mode=importlib"]
xfail_strict = true
log_cli_level = "warn"
pythonpath = "src"
testpaths = ["tests"]

[tool.black]
exclude = '''
/(
    \.git
  | \.tox
)/
'''
skip-magic-trailing-comma = true

[tool.isort]
profile = 'black'
skip_gitignore = true

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
exclude_also = [
  "if TYPE_CHECKING:",
  "if __name__ == .__main__.:",
  "except ImportError",
]

[tool.ruff]
# line-length = 88
line-length = 130
lint.select = [
  "E",
  "F",
  "W",   # flake8
  "B",   # flake8-bugbear
  "I",   # isort
  "ARG", # flake8-unused-arguments
  "C4",  # flake8-comprehensions
  "EM",  # flake8-errmsg
  "ICN", # flake8-import-conventions
  "ISC", # flake8-implicit-str-concat
  "G",   # flake8-logging-format
  "PGH", # pygrep-hooks
  "PIE", # flake8-pie
  "PL",  # pylint
  "PT",  # flake8-pytest-style
  "PTH", # flake8-use-pathlib
  "RET", # flake8-return
  "RUF", # Ruff-specific
  "SIM", # flake8-simplify
  "T20", # flake8-print
  "UP",  # pyupgrade
  "YTT", # flake8-2020
  "EXE", # flake8-executable
  "NPY", # NumPy specific rules
  "PD",  # pandas-vet
]
lint.extend-ignore = [
  "PLR",    # Design related pylint codes
  "E501",   # Line too long
  "B028",   # No explicit stacklevel
  "EM101",  # Exception must not use a string literal
  "EM102",  # Exception must not use an f-string literal
  "G004",   # f-string in Logging statement
  "RUF015", # Prefer next(iter())
  "RET505", # Unnecessary `elif` after `return`
]
lint.typing-modules = ["mypackage._compat.typing"]
lint.unfixable = [
  "T20",  # Removes print statements
  "F841", # Removes unused variables
]
lint.flake8-unused-arguments.ignore-variadic-names = true
lint.pydocstyle.convention = "numpy"
lint.isort.known-first-party = ["template"]
src = ["src"]
exclude = []
extend-exclude = ["tests"]

[tool.docformatter]
pre-summary-newline = true
black = true

[tool.mypy]
ignore_missing_imports = true
namespace_packages = true
explicit_package_bases = true
no_implicit_optional = false
disable_error_code = "annotation-unchecked"

[dependency-groups]
dev = [
    "pyside6>=6.9.1",
]

[project.scripts]
